Fixed comments in git
